MdeModulePkg: Fix potential overflow issue in HiiLib.c and ConfigRouting.c
Fixes a potential overflow in the comparison for the block data offsets
using UINT16 by casting to UINT32 before addition to prevent overflow.

+++ MdeModulePkg/Library/UefiHiiLib/HiiLib.c
@@ -2120,8 +2120,8 @@ GetBlockDataInfo (
   while ((Link != &BlockArray->Entry) && (Link->ForwardLink != &BlockArray->Entry)) {
     BlockData    = BASE_CR (Link, IFR_BLOCK_DATA, Entry);
     NewBlockData = BASE_CR (Link->ForwardLink, IFR_BLOCK_DATA, Entry);
-    if ((NewBlockData->Offset >= BlockData->Offset) && (NewBlockData->Offset <= (BlockData->Offset + BlockData->Width))) {
-      if ((NewBlockData->Offset + NewBlockData->Width) > (BlockData->Offset + BlockData->Width)) {
+    if ((NewBlockData->Offset >= BlockData->Offset) && (NewBlockData->Offset <= ((UINT32)BlockData->Offset + BlockData->Width))) { // MU_CHANGE: CodeQL
+      if (((UINT32)NewBlockData->Offset + NewBlockData->Width) > ((UINT32)BlockData->Offset + BlockData->Width)) { // MU_CHANGE: CodeQL
         BlockData->Width = (UINT16)(NewBlockData->Offset + NewBlockData->Width - BlockData->Offset);
       }
